cadastro do historico de consumo com valores em porcentagem(%)

+++ src/main/java/sistemaCaptura/log/metodos/Logs.java
@@ -20,7 +20,7 @@ public class Logs {
     private static final int LIMITE_CPU = 60;  // Defina o limite máximo de CPU conforme necessário
     private static final int LIMITE_RAM = 60;  // Defina o limite máximo de RAM conforme necessário
 
-    public static void gerarLog(Maquina maquina, Long consumoCpu, Long consumoRam, Long consumoDisco) {
+    public static void gerarLog(Maquina maquina, Long consumoCpu, double consumoRam, double consumoDisco) {
         LocalDate dataAtual = LocalDate.now();
         DateTimeFormatter dateFormat = DateTimeFormatter.ofPattern("yyyyMMdd", Locale.ENGLISH);
         String nomeArquivo = dateFormat.format(dataAtual) + "_log.txt";
@@ -33,7 +33,7 @@ public class Logs {
         }
     }
 
-    private static void adicionarMensagens(String caminhoCompleto, LocalDate dataAtual, Maquina maquina, Long consumoCpu, Long consumoRam, Long consumoDisco) {
+    private static void adicionarMensagens(String caminhoCompleto, LocalDate dataAtual, Maquina maquina, Long consumoCpu, double consumoRam, double consumoDisco) {
         try (BufferedWriter writer = Files.newBufferedWriter(Path.of(caminhoCompleto), StandardOpenOption.APPEND)) {
 
             String mensagemSuporte = "Suporte foi solicitado para arrumar  a maquina (" + maquina.getNome() + ").";
@@ -41,7 +41,7 @@ public class Logs {
             // Adicionar mensagem relacionada ao consumo máximo de CPU e RAM
             String mensagemConsumo = String.format("O consumo de CPU estourou o máximo sugerido (%d%%). O consumo de RAM atingiu o máximo sugerido (%d%%) de acordo com o nome da máquina.%n", LIMITE_CPU, LIMITE_RAM);
 
-            String dados = String.format("Data/Hora: %s%nNome da Máquina: %s%nConsumo CPU: %d%%%nConsumo RAM: %d bytes%nConsumo Disco: %d GB%nMensagem para Suporte: %s%n%s%n",
+            String dados = String.format("Data/Hora: %s%nNome da Máquina: %s%nConsumo CPU: %d%%%nConsumo RAM: %.2f bytes%nConsumo Disco: %.2f GB%nMensagem para Suporte: %s%n%s%n",
                     dataAtual, maquina.getNome(), consumoCpu, consumoRam, consumoDisco, mensagemSuporte, mensagemConsumo);
 
             writer.write(dados);
